feat: file to call creation new seeders

+++ src/shared/seeders/scriptSeeders/index.js
@@ -9,12 +9,21 @@ const rl = readline.createInterface({
 rl.question('>> Insert a seeder name: ', function (answer) {
   const timestamp = Date.now();
   const fileName = `Seed${(answer.charAt(0).toUpperCase() + answer.slice(1)).replace(/\s/g, '')}${timestamp}.ts`;
+
+  if (fs.existsSync(`./src/shared/seeders/${fileName}`)) {
+    console.log('Failed to create seed. Seed already exists!');
+    return;
+  }
+
   fs.writeFileSync(`./src/shared/seeders/${fileName}`, '', err => {
     if (err) {
       console.log('Failed to create seed', err);
       throw err;
     }
-    console.log('Seed created successfully!');
   });
+
+  if (fs.existsSync(`./src/shared/seeders/${fileName}`)) {
+    console.log('Success to create seed');
+  }
   rl.close();
 });
